requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
alembic
python-dotenv
strawberry-graphql